requirements.txt
google-cloud-dialogflow
python-dotenv
python-telegram-bot
vk_api
requests